requirements.txt
# Direct Dependencies
fastapi==0.68.1
psycopg2==2.9.1
python-decouple==3.4
SQLAlchemy==1.4.23
sqlmodel==0.0.4
uvicorn==0.15.0

# Secondary Dependencies
asgiref==3.4.1
click==8.0.1
colorama==0.4.4
greenlet==1.1.1
h11==0.12.0
pydantic==1.8.2
sqlalchemy2-stubs==0.0.2a15
starlette==0.14.2
typing-extensions==3.10.0.2
